Move bughunt utility helpers into module
